isl_factorizer: keep track of basic set
Since the isl_factorizer is specific to a particular basic set,
it may as well keep track of this basic set.
This allows the isl_factorizer_every_factor_basic_set function
introduced in the next commit to read off the basic set
instead of requiring it to be specified as a separate argument.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>