interface: isl_class::methods: use actual method name as key
The field is already documented as being keyed on the method name,
but this originally referred to the name of the C function
from which the method is derived.
This was later adjusted to the C function name with type suffixes removed.
Use the actual method name as it will appear in the bindings instead.
This will make it easier to group methods that are originally
defined in different classes as they have the same method name,
but different type prefixes.

Since the order in which the groups of methods are considered
is determined by this key, the change in key causes some changes
in the order of the methods in the output.
In particular, function names with a "get_" infix have
this removed from the method name and may therefore end up
in a different position.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>