isl_constraint_get_bound: normalize returned affine expression
The motivation for this change is essentially the same as that of
isl-0.22-48-g8956a09004 (isl_aff_add: normalize result,
Fri Dec 2 08:48:52 2016 +0100).
That is, if the original set from which the constraint was obtained
has more local variables than those that appear in this particular
constraint (and therefore also in the extracted bound),
then they would still be present with a zero coefficient.
They would therefore not get printed, but would still be taken
into account when comparing the affine expression to
other affine expression.  In particular, a zero affine expression
would not necessarily be considered obviously equal
to some other zero affine expression (without the hidden local variables).
Normalize the affine expression such that at least some (the final)
of these zero coefficient local variables get removed.

This change leads to a minor regression in the AST generation results.
However, the fact that the originally generated code is slightly
better appears to be based more on luck than on design.
In particular, before this commit, the relevant affine expression
{ [[] -> [i0, i1, i2]] -> [(18 + floor((5 + i0)/14))] }
would have an extra hidden local variable floor((i0)/14).
While taking the gist of this expression with respect
to the build context (14*floor((i0)/14) <= -9 + i0),
floor((5 + i0)/14) would be discovered to be equal to 1 + floor((i0)/14)
and the affine expression would get rewritten in terms
of the (originally) hidden floor((i0)/14).
This relies on local variables getting rewritten in terms
of earlier local variables.  However, the current ordering
of local variables does not guarantee that an earlier
local variable is always "better" than a later local variable.
So while it would be possible to take into account the local
variables of the context and to try and rewrite later
local variables in the affine expression in terms of
earlier local variables in the context, this is not guaranteed
to produce better results in all cases.
The minor regression is therefore simply accepted at this point.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>