isl_basic_map_plain_unshifted_simple_hull: drop gauss calls
It is not clear why those calls were introduced in the first place,
but they may result in some of the integer divisions copied
from the other basic map getting simplified using the equality
constraints.  The integer division expressions in the two basic maps
may therefore no longer be the same.
In particular, a constraint involving the simplified integer division
may appear in the result, while it is not obvious that this constraint
also hold for the other basic map.
In practice, for a constraint involving such an integer division
to appear in the result, it would have to be implied by
one of the two div constraints as those are the only constraints
in the basic map to which the integer division was added that
involve that integer division.
This means only redundant constraints could be added to the result
in this way.  Even though this should be harmless, it is confusing and
it is better not to add such constraints to the result.
Drop the gauss calls to avoid them getting added.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>