isl_ast_expr_from_aff: special case test for even value
In particular, an expression of the form

	(f - 1) mod 2

tests whether f is even, i.e., it is equal to 1 when f is even.
Replace the expression by

	1 - (f mod 2)

Note that an expression that is already of this form is left untouched,
resulting in a (rather specialized) normalization,
making the generated AST expression less dependent
on the internal representation.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>