isl_set_min_multi_pw_aff: avoid domain split in some special cases
In particular, if one expression is always better than
(or equal to) the other on their shared domain, then pick
the first one to avoid the risk of special casing
that part of the domain where they are equal.

With A and B the cells of the two expressions and
C the set of elements where the expression associated to B
is better than the expression associated to A,
it would be sufficient to check that C includes both A and B.
However, this requires an unconditional computation of C.
This commit looks for the specific case where
A is a subset of B and B is a subset of C
(which guarantees that C includes both A and B) and
only computes C if A is indeed a subset of B,
avoiding the computation of C in cases where
the entire check is assumed to be unlikely to succeed.
This decision can be revisited if it turns out
that this assumption was incorrect.

Reviewed-by: Tianjiao Sun <tianjiao.sun@cerebras.net>
Signed-off-by: Sven Verdoolaege <sven@cerebras.net>