add isl_set_every_basic_set
Similarly to isl_union_map_every_map and isl_union_set_every_set,
this function checks whether the callback functions returns true
on each basic set.

This function will be used in the next commit.
Since it deals with basic sets,
it is not exposed to the user by default.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>