add exported isl_pw_multi_aff_preimage_domain_wrapped_domain_pw_multi_aff
This is a convenience function around
isl_pw_multi_aff_pullback_pw_multi_aff that allows plugging
in a function into part of the domain.
It is exported to complement the other pullback and preimage functions.

The name is unfortunately a bit long.
Since "pullback" is essentially the same as "preimage_domain",
an alternative name would be
isl_pw_multi_aff_pullback_wrapped_domain_pw_multi_aff,
but that does not sound quite right.
Note that in the bindings, both the object type and the argument type
get dropped from the method name, so this method name is still
quite manageable.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>