isl_map_simplify.c: uset_gist: do not mark input empty if not known
If the integer affine hulls of the input and the context
have an empty intersection, then an empty result can be returned,
but the input itself should not be marked as empty.

The test cases need to be constructed in a specific space
to ensure that the basic sets that reach uset_gist are
the same as those in the original input.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>