isl_set_get_simple_fixed_box_hull: do not compute integer divisions
The reason for the original call to isl_map_compute_divs
is presumably to ensure the set description of which the bounds
are investigated does not involve any unknown local variables.
However, the computation of explicit integer division representations
for the local variables can be costly on some inputs and can
cause a split up into many different pieces rendering
the subsequent steps expensive as well.
Note that this computation is unlikely to produce any useful
lower bounds, especially since bounds that involve integer divisions
are ignored.  Even if it does, using them would go against the intent
of isl_set_get_simple_fixed_box_hull, which is to only look
for bounds that are explicitly available in the input representation.

Drop any constraints that involve unknown local variables instead.
Do this only for the set that is used for iterating over
the constraints.  The set that is used to compute the corresponding size
does not need to have its unknown local variables removed.

The new test case is not one that would take a long time
without this change, but rather just one that involves
unknown local variables to ensure this function keeps
working for such inputs.

Reported-by: NieWang <4477588@qq.com>
Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>