set sysroot if clang version used for bindings generation cannot find stdio.h
For reasonably recent versions of clang, the clang driver
is used to determine the options to be used for parsing
the header file from which the bindings will be generated.
On modern version of OS X, however, mainline clang does not
appear to be set up correctly to find the standard header files.
In particular, they are not in /usr/include, since that
is completely missing.
The Xcode version of clang sets -isysroot to point
to those header files.  Try and do the same if clang
cannot find the header files by default.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>