AST printing: make explicit printing of outermost block optional
If the outermost node of an AST is a block node, then it always
gets printed as such.  It is not clear if this was the intention or
if this was an oversight.  Either way, the default cannot be changed now.

The extra braces corresponding to this outermost block node
can be quite distracting though, so add an option that allows
the printing of these braces to be skipped.

Note that isl_ast_node_print also gets called internally
from print_body_c, but in this case the node is known
not to be a block as otherwise need_block would have returned 1.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>