isl_set_get_simple_fixed_box_hull: eliminate unknown local variables
The set from which a lower bound is constructed is not allowed
to have unknown local variables.
Originally, unknown local variables got removed by computing
explicit representations as integer divisions.
This can be quite costly, however, and isl-0.22.1-195-g240551690c
(isl_set_get_simple_fixed_box_hull: do not compute integer divisions,
Sun Mar 22 23:42:07 2020 +0100) replaced this by a removal
of all constraints involving unknown local variables.
This is also more in line with the intent of
isl_set_get_simple_fixed_box_hull, which is to only look
for bounds that are explicitly available in the input representation.

Unfortunately, this can also produce less accurate results
in some cases, including the test case that was added in that commit.
Remove the unknown local variables using (Fourier-Motzkin) elimination
instead.  This is more expensive than simply dropping constraints,
but should still be a lot less expensive than computing
explicit representations, especially if the latter results
in a lot of different cases.
It does mean that constraints will be considered for lower bounds
that do not appear directly in the input representation, but
the additional constraints considered are simple combinations of
those that do.  Furthermore, users of this function in the language
bindings cannot perform this elimination themselves since
it is considered low-level and not exported to the bindings.
Doing the elimination beforehand would also mean
that the set that is used for computing the size is also affected,
which is not the case when doing it internally.
It is not clear if this has any practical effect, however.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>