build extract_interface using build compiler
Since extract_interface is only used on the build system,
it should be built using the build compiler.
Inside the interface subdirectory, check for the build compiler and
use it throughout.  Do this instead of overriding the host system
to allow the caller of isl's configure to override
the host and build compilers separately.

The top level also needs to detect the build compiler
since EXEEXT may be different on the host and the build system.

It is not entirely clear if this is the right approach.
In particular, it is not clear how libtool is supposed
to be used in conjunction with a build compiler.
An alternative would be to have the top level
pass in the top level build compiler as
the (host) compiler for the interface subdirectory.
However, this needs special care to ensure that
the top level is not affected and would still affect
all subdir configures.  Even if there is currently only one,
this might cause problems in the future.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>