introduce isl_stat_non_error_bool
There are quite a few places where a temporary isl_bool result
needs to be converted to an isl_stat.  Provide a designated function
for this conversion to save space and to clarify the intent
of these code fragments, similarly to isl_stat_non_null.
This commit only introduces the function.  Calls will be added later.

Even though this function is mainly meant to be used internally,
it is exposed to the user since isl_stat_non_null is exposed as well.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>
Reviewed-by: Tianjiao Sun <tianjiao.sun@cerebras.net>
Signed-off-by: Sven Verdoolaege <sven@cerebras.net>