test_inputs/codegen/pldi2012/figure8_b.in: specify "separate" option
The currently generated output depends on the order
in which the two statements are considered.
In particular, the call to isl_set_make_disjoint inside
compute_partial_domains produces different results
depending on this order.
The next commit changes the order in which maps are stored
inside a union map object and would therefore result
in a different output.

Specify the "separate" option to ensure the same output
is generated independently of the order.
Note that this matches the default behavior of CodeGen++
of "removing overhead from depth 1 loops".

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>