isl_*_involves_locals: use stricter definition
That is, consider an expression to involve local variables
as soon as any local variable is present, even
if it only appears with a zero coefficient.
In practice this should not make a difference since
local variables that only appear with zero coefficients
should get removed automatically.
If this somehow does not happen, then the new definition
of these functions may be more confusing, but it should be safer
as it is usually used in situations where it is important
that the number of local variables is zero over the entire expression.

The new version is also a bit more efficient
since it does not need to examine the coefficients.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>