m4/ax_prog_cc_for_build.m4: do not override host compiler dependency style
Without these changes, the dependency style of the build compiler
would override that of the host compiler.
The next commit will allow users to specify a build compiler
for building extract_interface, but everything else will
be built using the host compiler.
If the user does not need extract_interface, then they
should not be forced to specify a build compiler,
whereas they may want to specify a host compiler,
which may have a dependency style that is different from
that of the automatically detected build compiler.
The dependency style of the build compiler should
therefore not override that of the host compiler.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>