isl_pw_*_dup: ensure result is identical to input
The original implementation of isl_pw_*_dup calls isl_pw_*_add_piece,
which may not necessarily add the piece, in particular if the domain
turns out to be (obviously) empty.
Extract out a variant of isl_pw_*_add_piece that always adds the piece
and use that in isl_pw_*_dup instead.

Now, an expression with an obviously empty domain should arguably
not have been added in the first place, but the emptiness
may have only been discovered after the piece was added.
Having isl_pw_*_dup return an expression with fewer cells
can cause confusion, especially when using isl_pw_*_restore_base_at,
which may call isl_pw_*_dup internally, but is expected
to simply replace a base expression and not to let the given
or any other cell disappear.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>