isl_basic_map_reduce_coefficients: extract out reduce_coefficients
This will allow this part of isl_basic_map_reduce_coefficients
to be reused in an upcoming commit.

The extracted function is responsible for marking the basic map
as having had its coefficients reduced to ensure that it doesn't
get marked after some possible further constraint manipulations
after the call to reduce_coefficients.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>