isl_basic_map_reduce_coefficients: reduce again after gauss
Much of the effect of the reduction may be undone by a call to
isl_basic_map_gauss.  Even though the call to isl_basic_map_gauss
itself may be useful to exploit any newly discovered equality
constraint, the coefficients should be reduced again
to ensure the function does what it is meant to do.

In particular, this allows the first of the newly added test cases
to be coalesced.  The second test case checks that
the second reduction is not done after the removal
of any local variables.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>