isl_map_coalesce: reduce coefficients of fused intermediate results
Any coefficient reduction performed by the top-level isl_map_coalesce
is likely to be undone by the call to isl_basic_map_gauss
on the fused intermediate results.
Apply coefficient reduction again on the result to improve
the chance that the fused intermediate results can be coalesced
with other basic maps, as in the newly added test case.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>