clear ISL_BASIC_MAP_REDUCED_COEFFICIENTS when needed
That is, clear the flag when the coefficients (other than
the constant term) of an inequality constraint may get modified or
when a new inequality constraint is added.
Do so instead of clearing the flag every time the basic map
is changed in any way or when some progress is made during simplification.

This should more accurately reflect when the coefficients
may no longer be reduced and is more in line with how
the other flags are handled.

One case that was missed before in particular was the call
to isl_basic_map_gauss from within isl_basic_map_reduce_coefficients
itself.  It this call takes place, then the coefficients
may very well no longer be reduced, as evidenced by an upcoming commit.

Note that accurately keeping track of this flag is not all that important.
Since isl_basic_map_reduce_coefficients may leave the output
in an inconsistent state, it always needs to be followed
by a call to isl_basic_map_gauss, which would typically remove the flag.
The chance that the flag is still set on the next call
to isl_basic_map_reduce_coefficients is therefore very small.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>