isl_union_*: ignore parameters when looking up entries
That is, once an entry has been found with the right (tuple) hash value,
ignore the parameters when checking whether the hash values correspond
to the same space.
This complements the move to using the tuple hash in the hash table and
is valid for the same reason.
With both changes in, look-ups can be performed
without first aligning parameters.

Note that union expressions that contain at most a single
base expression per domain space (i.e., those defined in isl_union_single.c)
already ignore parameters when looking up entries and
therefore do not require any changes.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>