isl_aff.c: pw_multi_aff_from_map_div: allow nested integer divisions
That is, allow the constraint from which an integer division
is constructed to itself involve integer divisions.
This prepares for an upcoming commit that will extend the detection
of the constraint in isl_basic_map_find_output_lower_div_constraint
in a similar way.

Note that pw_multi_aff_from_map_div now uses the local space
of the basic map, which may involve local variables
without a known integer division expression, even though
the coefficients of those variables are known to be zero
in the constraint from which the affine expression is constructed.
This still causes problems since such local variables
are not expected to appear in the local space at all.
Remove these local variables before the construction
of the affine expression.

Reviewed-by: Tianjiao Sun <tianjiao.sun@cerebras.net>
Signed-off-by: Sven Verdoolaege <sven@cerebras.net>