isl_basic_map_find_output_upper_div_constraint: allow integer divisions
That is, allow the constraint to refer to local variables
with a known expression as an integer division, as long
as this integer division does not involve any output variables.

This allows more types of functions to be directly read off
from the constraints when converting a binary relation to a function,
without having to resort to parametric integer programming.

In particular, this ensures that the new test case is obviously equal
to the original description when first parsed as a binary relation and
then converted to a function, while parametric integer programming
may result in a different expression (and currently does).

Reviewed-by: Tianjiao Sun <tianjiao.sun@cerebras.net>
Signed-off-by: Sven Verdoolaege <sven@cerebras.net>