export isl_aff_is_cst
This allows checking whether an affine function is constant.

Note that the similarly named isl_pw_aff_is_cst and
isl_multi_pw_aff_is_cst are not exported because
their names are misleading.  In particular, they only
check if an expression is piecewise constant,
which is not clear from their names.

Signed-off-by: Tianjiao Sun <tianjiao.sun@cerebras.net>
Signed-off-by: Sven Verdoolaege <sven@cerebras.net>