isl_basic_map_drop_redundant_divs_ineq: avoid possible spooky action at a distance
In particular, when looking for integer divisions that involve
some other integer division, ignore local variables
that have no known expression as an integer division.
The coefficients of these local variables may not always
get updated because they are irrelevant.
This means that the other integer division may incorrectly
get detected as being involved in some other integer division.
This in itself is relatively harmless since at worst
it results in a redundant integer division not getting eliminated.
However, this may have consequences further down the line
in operations that depend on the internal representation.
Most importantly, this may depend on the state of memory
that got reused from a completely unrelated isl object,
possibly resulting in other output depending
on operations performed on completely different objects.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>