isl_pw_qpolynomial_bound: handle differences in domain dimension names
If the domain of the input polynomial is a wrapped relation,
then internally, the domain dimensions of this wrapped relation
are temporarily treated as a sequence of parameters.
They therefore need to have the same names in both the polynomial
expression itself and its domain as otherwise
subsequent operations may complain about non-matching spaces.
These names can be different in particular if an initial
universe domain is intersected with a set where these names
are different.

Reset the space of the domain to the domain space
of the polynomial to ensure the names are the same.

Reported-by: Nuno Miguel Machado Nobre <nunomiguelnobre@gmail.com>
Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>