interface: introduce Method class
A couple of cpp_generator methods take both an isl_class and a FunctionDecl
argument, representing (in reverse order) the isl function from
which a method is derived and the class to which it belongs.
Replace the two arguments by a single argument by combining
the two in a Method object.
The static function is_subclass_mutator is more naturally considered
to be a method of the new Method class.
The other cpp_generator methods modified by this commit
end up calling one or more of the above methods.

The Method class will be further extended in subsequent commits.

Note that cpp_generator::is_implicit_conversion
is not turned into a method of Method because it relies
on cpp_generator::is_subclass, which needs access to the classes.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>