interface: ConversionMethod: support conversion of "this"
That is, keep track of the type to which "this" should be converted and
perform the conversion (if needed) in the call to the original
method from the conversion method.
This type is currently always the type of the class, indicating
that "this" should not be converted, but this commit prepares
for conversion methods that do convert "this".

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>