interface: handle automatic argument conversion through ConversionMethod subclass
That is, instead of passing along an extra "convert" argument,
store the required information in a subclass of Method.
Printing out the implementation is already handled by two separate methods.
Before, the extra argument was used to pick the right method.
Now, the type of method (ConversionMethod or just Method) determines
which printing method gets called.
Printing out the header is done by a single method.
This used to call cpp_generator::class_printer::get_param with
the "convert" argument and now calls a get_param method of Method
that is overridden by ConversionMethod.
The latter then ends up calling cpp_generator::class_printer::get_param,
but this method is no longer called for a non-conversion method,
so it no longer needs to handle an empty "convert".

Note that the virtual print_method method is called
from a generic cpp_generator::class_printer::print_method_variants.
Only the cpp_generator::impl_printer versions have a different behavior
depending on whether a Method or a ConversionMethod gets printed.
The cpp_generator::decl_printer versions are now identical.

Also note that ConversionMethod::get_param does not call
cpp_generator::class_printer::get_param directly, but instead calls
a lambda passed in during construction.
This means ConversionMethod itself does not need to be aware
of how the argument conversion is performed and does not
store "convert" directly.
This will be useful when ConversionMethod is reused to convert
the "this" argument and not any of the other arguments.
This future reuse of ConversionMethod is also the main motivation
for introducing the ConversionMethod class.

Since ConversionMethod does not store "convert" directly,
the printing of the implementations uses an indirect way
of figuring out whether an argument is converted.
In particular, it checks if the method argument is different
from the C function argument.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>