interface: pass isl type arguments as const reference to conversion methods
Unlike methods that call an isl C function,
a conversion method never calls release() on an isl type argument,
so they can all be passed as const references.

This does not have any effect on any of the currently exported functions
because the methods with automatic conversion do not take
any isl type arguments.
However, when ConversionMethod gets reused to convert
the "this" argument in an upcoming commit, most
of the other arguments will be of isl type and
these should be passed as const references.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>