interface: Method: keep track of method kind
Keep track of the method kind inside the Method object
such that it does not need to be passed along separately.
In fact, provide a constructor that determines
this kind automatically.
Since the cpp_generator::get_method_kind method for determining
this kind is now only used from within this constructor,
it is turned into a static funtion (and renamed to get_kind).
Similarly, the function_kind enum is moved to Method.

Some methods that take a function declaration and a method kind
are also changed to take a Method object instead.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>