C++ bindings: stop creating include/isl/cpp directory
The creation of this directory was introduced by
isl-0.18-1066-gd5ca2343f5 (introduce (unchecked) cpp bindings
generation, Tue Aug 1 17:09:54 2017 +0200),
but it is not needed since the header is created inside include/isl and
not inside include/isl/cpp.
This may have been a remnant of some earlier version of the commit.

Continue creating include/isl, even if it is not strictly required
because include/isl/stdint.h is already placed there.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>