move print_{named_,}method_header to cpp_generator::class_printer
This allows the methods to reuse the fields of
cpp_generator::class_printer.  This includes the new
"declarations" field that replaces the "is_declaration"
argument.  Every class_printer object is either
a decl_printer or an impl_printer, which consistently
print declarations and definitions respectively,
such that there is no need to pass this information
on every call to print_method_header.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>