re-export isl_multi_*_zero as isl_space_zero_multi_*
In the bindings, it can be more convenient to use a method
on the space in which the object lives than to call a named constructor
with the space as argument.

Type based suffixes usually refer to argument types.
In this case, the suffix refers to the result type.
This is similar to the functions
isl_{set,map}_{lexmin,lexmax}_pw_multi_aff.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>