isl_{set,map}_{lower,upper}_bound_multi_pw_aff: take into account explicit domain
When isl_map_{lower,upper}_bound_multi_pw_aff is called
on an isl_multi_pw_aff that is not zero-dimensional,
then the domain of the map gets intersected with the domains
of the entries of the isl_multi_pw_aff.
It therefore seems natural that in case of a zero-dimensional
isl_multi_pw_aff, the domain of the map should also
get intersected with the explicit domain of the isl_multi_pw_aff.
Otherwise, calling this function on a zero-dimensional
isl_multi_pw_aff would have no effect.

Taking into account the explicit domain will also turn out
to be useful for testing the fix to
isl_map_{min,max}_multi_pw_aff in the next commit.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>