re-export isl_multi_*_identity_on_domain_space as isl_space_*_multi_*
In the bindings, it can be more convenient to use a method
on the space in which the domain of the object lives than
to call a named constructor with the space as argument.

Type based suffixes usually refer to argument types.
In this case, the suffix refers to the result type.
It still comes before "on_domain", though,
since it qualifies "identity" and not "domain".

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>