re-export isl_multi_aff_multi_val_on_domain_space as isl_space_multi_aff_*
In the bindings, it can be more convenient to use a method
on the space in which the domain of the object lives than
to call a named constructor with the space as argument.

Sensible names for this new better exportable function are
isl_space_multi_val_on_domain_multi_aff,
where the suffix refers to the result type, and
isl_space_multi_aff_on_domain_multi_val,
where the suffix refers to the only remaining argument type.

The second name is preferred because it results in a shorter name
in the bindings, without losing information.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>