AST generation: check that partial schedule of band node is total on instances
That is, check that it does not filter out any instances.

Filtering out instances should be performed by filter nodes.
Band nodes, on the other hand, should not filter out any instances.
Ideally, this property should be checked during construction,
but it may be fairly costly and non-trivial, especially
in the presence of extension nodes.
At least perform the check during AST generation
to prevent the mysterious disappearance of statement instances.
Do this in build_ast_from_band, which already checks that
the band node does not introduce any new parameters.

Requested-by: Tomofumi Yuki <tomofumi.yuki@gmail.com>
Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>