cpp_generator::param2cpp: extract out cpp_type_printer
This separates out the difference in type printing for
the regular and the checked C++ interface.
The cpp_type_printer will also be reused
for specializing the printing of types
for the templated C++ interface.

The cpp_type_printer::isl_type method is extracted out
at the same time to make the implementation of
cpp_type_printer::param more consistent over all
different types.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>