plain_cpp_generator: copy methods inherited from superclasses in constructor
The next commit will extract a generic cpp_generator
from plain_cpp_generator.  The copy_super_methods
will be moved to the generic cpp_generator because
it will also be useful for the templated C++ interface
that will be introduced later.
The call to copy_super_methods is first moved
to the (now) plain_cpp_generator constructor,
so that it can be moved to the cpp_generator constructor
along with the moved of the copy_super_methods itself.
This means it will get called automatically
for the templated C++ interface when it gets introduced.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>