plain_cpp_generator::class_printer: extract out plain_printer
Some methods in class_printer depend on the "checked" field
(possibly indirectly through the type printer).
This field will not be moved to cpp_generator when
it gets extracted out from plain_cpp_generator,
while class_printer will get moved to cpp_generator.
These methods therefore first need to be put
in an intermediate plain_printer.
The plain_printer keeps track of its own reference
to the generator because the one in class_printer
will become a reference to a generic cpp_generator.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>