export isl_aff_param_on_domain_space_id as isl_space_param_aff_on_domain_id
In the bindings, it can be more convenient to use a method
on the space in which the domain of the object lives than
to call a named constructor with the space as argument.

The "id" suffix is kept because a "name" variant might be added later.
The "aff" suffix is then best places before "on_domain"
as a qualifier on "param".

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>