isl_mat_alloc: avoid undefined null pointer arithmetic
Technically, arithmetic on null pointers is undefined behavior,
even if only adding zero.  If isl_mat_alloc is called with
zero n_col argument, but nonzero n_row, the allocated isl_blk size is
still zero and isl_blk_alloc returns an empty allocation with the data
member set to NULL. Because n_row is nonzero, it is still used to
initialize the isl_mat row addresses using an offset to the NULL isl_blk
data member.

This is fixed by initializing the row array with NULL using calloc and
only conditionally assigning the pointer elements to an offset derived
from the isl_blk allocation.

Signed-off-by: Michael Kruse <isl@meinersbur.de>
Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>