isl_qpolynomial_bound_on_domain_bernstein: special case all constant polynomials
The base case of bernstein expansion would already special case
polynomials over a zero-dimensional domain and zero polynomials,
but this can be extended to all polynomials that are constant
expressions in the domain variables.
The test for a zero-dimensional domain is kept,
even if it corresponds to a special case of a constant polynomial.

Note that this makes the recent change to is_tight
to handle zero-degree polynomials redundant since this function
will now no longer get called in case of
zero-degree polynomials.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>