interface: rename one cpp_generator::type2cpp overload to param2cpp
There are three overloads of cpp_generator::type2cpp,
one taking a string representation of a C isl type,
one taking an isl_class and one taking a clang::QualType.
The first two both take a C isl type (as a string or
as an isl_class), but the third tries to convert
any parameter or return type.
Letting it have the same name as the other two is therefore confusing.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>