merge test_plain_unshifted_simple_hull and test_unshifted_simple_hull
These two testing functions are identical apart from
the specific simple hull function that is being called.
Specifying this function in the input data allows
the two testing functions to be merged into a single testing function.
This will make it easier to add another simple hull test
in the next commit.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>