isl_basic_map_align_divs: do not align against modified source
In many places, the code tacitly assumes that no integer division
appears before any other integer division on which it depends.
The swapping performed by isl_basic_map_align_divs may result
in integer divisions appearing in the wrong order.
This appears to have originally been handled
in isl_basic_map_align_divs by ordering the integer divisions
of the input.  As such, this does not guarantee that the integer divisions
of the output are ordered, but basic maps are usually aligned
with respect to each other and so the second call would
ensure that all integer divisions appear in an appropriate order.
This was broken by isl-0.20-13-gf64e1c0fef (isl_basic_map_align_divs:
improve error handling, Thu Aug 16 12:41:04 2018 +0200) because
it would perform the ordering of the input integer divisions
on a copy of the input.  The output integer divisions would
therefore be aligned to the ordered input integer divisions,
which may be different from their original order.
This could cause all kinds of problems on operations
that require aligned integer divisions, such as in the new simple hull
test case derived from a reported AST generation issue.

Fix the issue by ordering the output integer divisions
instead of the input integer divisions.
This assumes that the input integer divisions are already
ordered (or that they will get ordered by a second call
to isl_basic_map_align_divs), but this should be the case.
Still, for good measure, some extra calls to
isl_basic_map_order_divs are added before calls
to isl_basic_map_align_divs.

Reported-by: Michael Kruse <MichaelKruse@meinersbur.de>
Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>