move cpp_generator::print_downcast_* to cpp_generator::*_printer
This allows the methods to reuse the fields of
the parent cpp_generator::class_printer.
The "_decl" and "_impl" suffixes are removed because the methods now appear
in decl_printer or impl_printer.

cpp_generator::print_isa_type_template is an integral part
of cpp_generator::print_downcast_decl and is moved along with it.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>