isl_pw_multi_aff_substitute: only allow input dimension substitution
This is the only case that is currently used and
this restriction will make it easier to implement isl_set_substitute
in terms of isl_set_preimage_multi_aff.

Note that this also fixes the type argument passed to isl_set_substitute
from isl_dim_in to isl_dim_set, but this does not have any effect
in practice.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>