re-export isl_pw_aff_param_on_domain_id as isl_set_param_pw_aff_on_domain_id
In the bindings, it can be more convenient to use a method
on the domain of the object than
to call a named constructor with the set as argument.

The naming is analogous to isl_space_param_aff_on_domain_id,
which ensures that the new function is also exported
to the templated interface.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>