isl_pw_*_reset_space_and_domain: use isl_pw_*_restore_space
This reduces the dependence on the internal representation.

Note that even though a call to isl_pw_*_take_space
always needs to be followed by a corresponding call
to isl_pw_*_restore_space, it is perfectly fine
to call isl_pw_*_restore_space without a preceding call
to isl_pw_*_take_space.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>