isl_pw_*_set_dim_id: only modify space once
In particular, isl_pw_*_reset_space already modifies
the space of the piecewise object itself (together
with the spaces of the pieces), so there is no need
for isl_pw_*_set_dim_id to modify the space prior
to the call to isl_pw_*_reset_space.
Simply pass in the modified space.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>