isl_pw_*_gist_last: use isl_pw_*_{take,restore}_base_at
This reduces the dependence on the internal representation.

There is another access to FIELD to remove the other pieces,
but this one is not replaced by a call
to isl_pw_*_take_base_at because there is no corresponding
isl_pw_*_restore_base_at.
If the piecewise expression is known to have only one reference,
which is the case here, then calling isl_pw_*_take_base_at
still works, but it is too error-prone.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>