extract out shared isl_pw_*_{take,restore}_domain_at
This reduces the dependence on the internal representation.

Note that the calls to isl_pw_*_cow can be removed
because it is called by isl_pw_*_restore_domain_at
(or isl_pw_*_restore_base_at) when needed.
In isl_pw_*_gist_fn, the call cannot be removed
because this function also performs other manipulations
that require the object to have a single reference.

Also note that even though a call to isl_pw_*_take_domain_at
always needs to be followed by a corresponding call
to isl_pw_*_restore_domain_at, it is perfectly fine
to call isl_pw_*_restore_domain_at without a preceding call
to isl_pw_*_take_domain_at.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>