isl_space.c: copy_ids: free original identifier
The function set_id simply overwrites the stored identifier,
leaving it up to the caller to free the originally stored
identifier (if any).
The function copy_id is currently always called
in a situation where no identifier has already been stored.
This will change in the next commit.
Free the original identifier, which has no effect
on all current calls.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>