isl_pw_*_exploit_equalities_and_remove_if_empty: use isl_pw_*_*_base_at
This reduces the dependence on the internal representation.

Furthermore, this is safer to use in case there are multiple references
to the modified piecewise expression.
isl_pw_*_exploit_equalities_and_remove_if_empty is currently only
called in cases where the piecewise expression is known
to only have a single reference.
However, this may change in the next commit.  At least,
after the next commit, it will no longer be obvious
that this function is only called on objects with a single reference.

Note that if this function should ever be called on an object
with multiple reference, then it is safe to modify all references
since exploiting equalities does not change the meaning
of the object.  A special isl_pw_*_restore_base_at_inplace
variant of isl_pw_*_restore_base_at is therefore used
that does not ensure that only a single reference is modified.

Reviewed-by: Tianjiao Sun <tianjiao.sun@cerebras.net>
Signed-off-by: Sven Verdoolaege <sven@cerebras.net>