isl_ast_codegen.c: generate_domain: construct call from non-gisted map
Using a gisted map removes information that could be used
to simplify expressions.
This is especially useful for the codegen/separate2.c test case
that is currently only able to obtain a relatively simple expression
because some unused integer division is still available,
but this will be removed in an upcoming commit.

Back in isl-0.11.1-7-g076a978c43 (isl_ast_codegen.c: generate_domain:
avoid infinite recursion, Tue Feb 19 16:46:33 2013 +0100),
the additional information had the opposite effect,
where using the non-gisted map would result in some test cases
being changed to include some expressions that are more complicated.
At this stage, the fall-out is minimal.

Note that the call to isl_map_coalesce is kept
to avoid splitting up the domain on some inputs, notably expansion.st.

Also note that this commit removes the only use of the single_valued
property in isl_ast_build.  The attribute itself will be removed
in the next commit.

Furthermore, all calls to add_domain now pass two copies
of the original inverse schedule.  One of the copies will be removed
in an upcoming commit.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>