isl_ast_build: stop keeping track of single-valuedness
This mechanism was introduced in isl-0.11.1-7-g076a978c43
(isl_ast_codegen.c: generate_domain: avoid infinite recursion,
Tue Feb 19 16:46:33 2013 +0100) to keep track of the fact
that the ungisted inverse schedule is necessarily single-valued.
Since the gisting of the inverse schedule was removed
in the previous commit, this property was no longer inspected and
so it can be removed completely.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>